requirements.txt
-e git+https://github.com/PetrDlouhy/django.git@view_permission#egg=Django
django-import-export
-e git+https://github.com/digi604/django-smart-selects.git#egg=django-smart-selects
psycopg2
python-memcached
-e git+https://github.com/auto-mat/django-remote.git#egg=django-remote
django-chart-tools
django-extensions
django-debug-toolbar
django-softhyphen==0.15
django-admin-enhancer
django-polymorphic
django-composite-field
django-registration==0.8
-e git+https://github.com/silverfix/django-nested-inlines.git#egg=django-nested-inlines
django-su
unidecode
createsend==2.6.0
ordereddict
-e git+https://github.com/PetrDlouhy/django-mass-edit#egg=django-mass-edit
-e git+https://github.com/PetrDlouhy/django-author.git#egg=django-author
-e git+https://github.com/PetrDlouhy/django-fieldsignals.git#egg=django-fieldsignals
django-timelog
django-admin-sortable2
reportlab
svg2rlg
flake8
modulus11
-e hg+https://bitbucket.org/mjs7231/django-dbbackup#egg=django-dbbackup
django-localeurl
pillow
-e git+https://github.com/PetrDlouhy/InvoiceGenerator.git#egg=InvoiceGenerator
-e git+https://github.com/PetrDlouhy/django-related-admin#egg=django-related-admin
newrelic
easy_thumbnails
django-crispy-forms
django-class-based-auth-views
django-subdomains
django-adminfilters
django-compressor
slumber
django-bleach
django-analytical
gunicorn
django-cache-utils
django-settings-context-processor
django-rosetta
django-oauth-toolkit
djangorestframework
